Move code into lib.rs, avoiding `ptp::ptp::` paths.
